Finishing base refactoring

+++ src/project/project.service.ts
@@ -1,6 +1,6 @@
 import { Repository } from 'typeorm';
-import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
+import { Injectable, NotFoundException } from '@nestjs/common';
 
 import { Project } from './entities/project.entity';
 import { CreateProjectDto } from './dto/create-project.dto';
@@ -11,7 +11,7 @@ export class ProjectService {
 
   constructor(
     @InjectRepository(Project) 
-    private repository: Repository<Project>,
+    private repository: Repository<Project>
   ) {}
 
   async create(createProjectDto: CreateProjectDto): Promise<Project> {
@@ -27,18 +27,38 @@ export class ProjectService {
   }
 
   async findOne(id: string): Promise<Project> {
-    return await this.repository.findOne(id); 
+    const project = await this.repository.findOne(id);
+
+    if (!project) {
+      throw new NotFoundException(`Project with id ${id} does not exists.`);
+    }
+
+    return project;
   }
 
   async update(id: string, dto: UpdateProjectDto): Promise<Project> {
     const project = await this.repository.findOne(id);
+
+    if (!project) {
+      throw new NotFoundException(`Project with id ${id} does not exists.`);
+    }
+    
     const merge = this.repository.merge(project, dto);
     return await this.repository.save(merge); 
   }
 
   async remove(id: string): Promise<string>  {
     const project = await this.repository.findOne(id);
-    this.repository.delete(project.idProject);
-    return `Project ${id} has been deleted`; 
+
+    if (!project) {
+      throw new NotFoundException(`Project with id ${id} does not exists.`);
+    }
+
+    try {
+      this.repository.delete(project.idProject);
+      return `Project ${id} has been deleted`; 
+    } catch (error) {
+      throw new Error(`Error Deleting Project! \nMessage: ${error}`);
+    }
   }
 }
